requirements.txt
pyOpenSSL==23.1.1
websockets==11.0.2
